+++ src/Oro/Bundle/OrderBundle/DependencyInjection/OroOrderExtension.php
@@ -19,6 +19,7 @@ class OroOrderExtension extends Extension
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
         $loader->load('services.yml');
         $loader->load('services_api.yml');
+        $loader->load('services_pdf_document.yml');
         $loader->load('form_types.yml');
         $loader->load('block_types.yml');
         $loader->load('controllers.yml');